propogate all parent detail variables to children
